Background Skills Chosen Signed-off-by: Christopher Knarvik <x23chris@gmail.com>
